refactor: split DB queries into queries.ts for SonarCloud coverage
Move all DB query functions (fetch*) into server-only queries.ts.
fetchDeclarations.ts now contains only types and pure functions
(buildIndicators, buildIndicatorG, assembleDeclaration) which are
fully covered by unit tests. Tests mock queries.ts instead of
fetchDeclarations.ts, so the pure functions execute during tests.
