+++ packages/app/src/modules/export/fetchDeclarations.ts
@@ -1,65 +1,18 @@
-import { and, eq, gte, inArray, lt, or } from "drizzle-orm";
+// Types and pure functions for building the declaration JSON response.
+// DB queries are in queries.ts (server-only, not unit-testable).
 
-import { db } from "~/server/db";
-import {
-	companies,
-	cseOpinions,
-	declarationCategories,
-	declarations,
-	employeeCategories,
-	jobCategories,
-	users,
-} from "~/server/db/schema";
+// Re-export queries for route handler convenience
+export {
+	fetchCategoriesByDeclaration,
+	fetchCseOpinionsByDeclaration,
+	fetchIndicatorGByDeclaration,
+	fetchSubmittedDeclarations,
+} from "./queries";
+export type { DeclarationRow } from "./queries";
 
-// ── Main query ───────────────────────────────────────────────────────
+import type { DeclarationRow } from "./queries";
 
-type DeclarationRow = Awaited<
-	ReturnType<typeof fetchSubmittedDeclarations>
->[number];
-
-export async function fetchSubmittedDeclarations(
-	dateBegin: string,
-	dateEnd: string,
-) {
-	return db
-		.select({
-			siren: declarations.siren,
-			year: declarations.year,
-			status: declarations.status,
-			compliancePath: declarations.compliancePath,
-			totalWomen: declarations.totalWomen,
-			totalMen: declarations.totalMen,
-			secondDeclarationStatus: declarations.secondDeclarationStatus,
-			secondDeclReferencePeriodStart:
-				declarations.secondDeclReferencePeriodStart,
-			secondDeclReferencePeriodEnd: declarations.secondDeclReferencePeriodEnd,
-			createdAt: declarations.createdAt,
-			updatedAt: declarations.updatedAt,
-			declarationId: declarations.id,
-			companyName: companies.name,
-			workforce: companies.workforce,
-			nafCode: companies.nafCode,
-			address: companies.address,
-			hasCse: companies.hasCse,
-			declarantFirstName: users.firstName,
-			declarantLastName: users.lastName,
-			declarantEmail: users.email,
-			declarantPhone: users.phone,
-		})
-		.from(declarations)
-		.innerJoin(companies, eq(declarations.siren, companies.siren))
-		.innerJoin(users, eq(declarations.declarantId, users.id))
-		.where(
-			and(
-				eq(declarations.status, "submitted"),
-				// Timestamps are stored in UTC (withTimezone: true)
-				gte(declarations.updatedAt, new Date(`${dateBegin}T00:00:00Z`)),
-				lt(declarations.updatedAt, new Date(`${dateEnd}T00:00:00Z`)),
-			),
-		);
-}
-
-// ── Indicators A–F (declarationCategories, steps 2–4) ────────────────
+// ── Types ────────────────────────────────────────────────────────────
 
 export type CategoryRow = {
 	step: number;
@@ -71,46 +24,6 @@ export type CategoryRow = {
 	womenMedianValue: string | null;
 	menMedianValue: string | null;
 };
-
-export async function fetchCategoriesByDeclaration(
-	keys: Array<{ siren: string; year: number }>,
-): Promise<Map<string, CategoryRow[]>> {
-	if (keys.length === 0) return new Map();
-
-	const pairConditions = keys.map((k) =>
-		and(
-			eq(declarationCategories.siren, k.siren),
-			eq(declarationCategories.year, k.year),
-		),
-	);
-
-	const rows = await db
-		.select({
-			siren: declarationCategories.siren,
-			year: declarationCategories.year,
-			step: declarationCategories.step,
-			categoryName: declarationCategories.categoryName,
-			womenCount: declarationCategories.womenCount,
-			menCount: declarationCategories.menCount,
-			womenValue: declarationCategories.womenValue,
-			menValue: declarationCategories.menValue,
-			womenMedianValue: declarationCategories.womenMedianValue,
-			menMedianValue: declarationCategories.menMedianValue,
-		})
-		.from(declarationCategories)
-		.where(or(...pairConditions));
-
-	const map = new Map<string, CategoryRow[]>();
-	for (const row of rows) {
-		const key = `${row.siren}-${row.year}`;
-		const existing = map.get(key) ?? [];
-		existing.push(row);
-		map.set(key, existing);
-	}
-	return map;
-}
-
-// ── Indicator G (jobCategories + employeeCategories) ─────────────────
 
 export type IndicatorGEntry = {
 	categoryName: string;
@@ -128,81 +41,11 @@ export type IndicatorGEntry = {
 	hourlyVariableMen: string | null;
 };
 
-export async function fetchIndicatorGByDeclaration(
-	declarationIds: string[],
-): Promise<Map<string, IndicatorGEntry[]>> {
-	if (declarationIds.length === 0) return new Map();
-
-	const rows = await db
-		.select({
-			declarationId: jobCategories.declarationId,
-			categoryName: jobCategories.name,
-			detail: jobCategories.detail,
-			declarationType: employeeCategories.declarationType,
-			womenCount: employeeCategories.womenCount,
-			menCount: employeeCategories.menCount,
-			annualBaseWomen: employeeCategories.annualBaseWomen,
-			annualBaseMen: employeeCategories.annualBaseMen,
-			annualVariableWomen: employeeCategories.annualVariableWomen,
-			annualVariableMen: employeeCategories.annualVariableMen,
-			hourlyBaseWomen: employeeCategories.hourlyBaseWomen,
-			hourlyBaseMen: employeeCategories.hourlyBaseMen,
-			hourlyVariableWomen: employeeCategories.hourlyVariableWomen,
-			hourlyVariableMen: employeeCategories.hourlyVariableMen,
-		})
-		.from(jobCategories)
-		.innerJoin(
-			employeeCategories,
-			eq(jobCategories.id, employeeCategories.jobCategoryId),
-		)
-		.where(inArray(jobCategories.declarationId, declarationIds));
-
-	const map = new Map<string, IndicatorGEntry[]>();
-	for (const row of rows) {
-		const existing = map.get(row.declarationId) ?? [];
-		existing.push(row);
-		map.set(row.declarationId, existing);
-	}
-	return map;
-}
-
-// ── CSE opinions ─────────────────────────────────────────────────────
-
 export type CseRow = {
 	type: string;
 	opinion: string | null;
 	opinionDate: string | null;
 };
-
-export async function fetchCseOpinionsByDeclaration(
-	keys: Array<{ siren: string; year: number }>,
-): Promise<Map<string, CseRow[]>> {
-	if (keys.length === 0) return new Map();
-
-	const pairConditions = keys.map((k) =>
-		and(eq(cseOpinions.siren, k.siren), eq(cseOpinions.year, k.year)),
-	);
-
-	const rows = await db
-		.select({
-			siren: cseOpinions.siren,
-			year: cseOpinions.year,
-			type: cseOpinions.type,
-			opinion: cseOpinions.opinion,
-			opinionDate: cseOpinions.opinionDate,
-		})
-		.from(cseOpinions)
-		.where(or(...pairConditions));
-
-	const map = new Map<string, CseRow[]>();
-	for (const row of rows) {
-		const key = `${row.siren}-${row.year}`;
-		const existing = map.get(key) ?? [];
-		existing.push(row);
-		map.set(key, existing);
-	}
-	return map;
-}
 
 // ── Build indicators from raw rows ───────────────────────────────────
